% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddGaussianOutliers.R
\name{AddGaussianOutliers}
\alias{AddGaussianOutliers}
\title{Add outliers using gaussian noise}
\usage{
AddGaussianOutliers(
  SequencingTable,
  drawoption = 10^-5,
  feedback = FALSE,
  setseed = FALSE
)
}
\arguments{
\item{SequencingTable}{a read counts table with the transcripts in row and the samples in column.}

\item{drawoption}{the way to pick outliers, if the value is smaller than 1 it is the probability of observation gets noisy. If it is higher than one it is the number of observations to change.}

\item{feedback}{logical, add a data frame containing informations on the outliers.}

\item{setseed}{set the seed.}
}
\value{
A sequencing table with outliers. Outliers are made using the formula:
\ifelse{html}{
    \out{k<sub>ij</sub><sup>&#119977</sup> = k<sub>ij</sub> + &#119977(0,σ<sub>i</sub>)}}{
    \deqn{k_{ij}^{N}=k_{ij}+N(0,σ_{j}^{N})}}.
More detail in the ABEILLE paper.
}
\description{
Create a copy of a sequencing table with outliers using gaussian noise.
}
\examples{
SequencingTable <- ExampleAbeilleDataSet
TableWithOutliers <- AddGaussianOutliers(SequencingTable,
drawoption = 10^-4, feedback = TRUE, setseed = 5)
print(head(TableWithOutliers$table))
print(head(TableWithOutliers$feedback))
}
