% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddOutriderOutliers.R
\name{AddOutriderOutliers}
\alias{AddOutriderOutliers}
\title{Add outliers using OUTRIDER method}
\usage{
AddOutriderOutliers(
  SequencingTable,
  SizeFactor,
  Zmean = log(3),
  Zsd = log(1.6),
  drawoption = 10^-5,
  feedback = FALSE,
  setseed = FALSE
)
}
\arguments{
\item{SequencingTable}{a read counts table with the transcripts in row and the samples in column.}

\item{SizeFactor}{vector containing all the size factor. The size factor of each sample can be compute using the function SizeFactors, the function use the estimatation of DESeq2.}

\item{Zmean}{mean of the gaussian distribution used. Default log(3).}

\item{Zsd}{standard deviation of the gaussian distribution used. Default log(1.6).}

\item{drawoption}{the way to pick outliers, if the value is smaller than 1 it is the probability of observation gets noisy. If it is higher than one it is the number of observations to change.}

\item{feedback}{logical, add a data frame containing informations on the outliers.}

\item{setseed}{set the seed.}
}
\value{
A sequencing table with outliers. Outliers are drawn using the formula:
\ifelse{html}{
    \out{u<sub>ij</sub> = log2(k<sub>ij</sub>/s<sub>j</sub> + 1) and k<sub>ij</sub><sup>O</sup> = round(s<sub>j</sub>2<sup>μ<sub>i</sub><sup>u</sup>±exp(<i>&#119977</i>)σ<sub>i</sub><sup>u</sup></sup>)}}{
    \deqn{u_{ij}=log_2⁡(k_{ij}/s_{j} +1) and k_{ij}^{O}=round(s_{j} 2^{μ_{i}^{u}±exp(N) σ_{j}^{u}}}}.
More detail in the ABEILLE paper.
}
\description{
Create a copy of a sequencing table with outliers using OUTRIDER formula.
}
\examples{
SequencingTable <- ExampleAbeilleDataSet
anno_table <- data.frame(sampleID = colnames(ExampleAbeilleDataSet),
Sex = sample(c("Male", "Female"), 128, replace = TRUE), Age = sample(1:89, 128, replace = TRUE),
Cohort = rep("GTEx database",128))
sizefactor <- SizeFactors(ExampleAbeilleDataSet, anno_table)
TableWithOutliers <- AddOutriderOutliers(SequencingTable, sizefactor, Zmean = log(3),
Zsd = log(1.6), drawoption = 10^-4, feedback = TRUE, setseed = 5)
print(head(TableWithOutliers$table))
print(head(TableWithOutliers$feedback))
}
