% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PickOutliers.R
\name{ComputeAberrantScore}
\alias{ComputeAberrantScore}
\title{Function computing Aberrant Score based on the decision tree}
\usage{
ComputeAberrantScore(my_df)
}
\arguments{
\item{divergence_score}{data frame of the same size as the initial dataset. The z-score can be compute with the function divergence_score.}
}
\value{
Data frame with additional column containing Aberrant Score.
}
\description{
This function is call from the function PickOutliers
}
\section{}{
NA
}

\examples{

}
