% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyAGE.R
\name{ComputeAnomalyScore}
\alias{ComputeAnomalyScore}
\title{Function computing Anomaly Score based on divergence score and delta count.}
\usage{
ComputeAnomalyScore(my_df)
}
\arguments{
\item{my_df}{data frame containing divergence score and delta count columns. Divergence scores and delta counts can be computed with functions DivergenceScore and DeltaCount of ABEILLE package.}
}
\value{
Data frame with additional column containing anomaly scores.
}
\description{
This function is call from the function IdentifyAGE
}
\section{}{
NA
}

\examples{

}
