% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaCount.R
\name{DeltaCount}
\alias{DeltaCount}
\title{Compute the Delta Count}
\usage{
DeltaCount(SequencingTable, ReconstructedTable)
}
\arguments{
\item{SequencingTable}{a read counts table with the transcripts in row and the samples in column.}

\item{ReconstructedTable}{the reconstructed table generate by an autoencoder.}
}
\value{
All Delta Count in a matrix. The computation is the following one:
\ifelse{html}{
    \out{Δ<sub>ij</sub> = log2(k<sub>ij</sub> / μ<sub>i</sub><sup>k&#770)}}{
    \deqn{Δ_{ij} = log2(k_{ij} / μ&hat_{ij}^{k&hat})}}.
More detail in the ABEILLE paper.
}
\description{
Metric used to find AGE, this definition is inspired from the log2 fold-change is not the classic one.
}
\examples{
SequencingTable <- ExampleAbeilleDataSet
ReconstructedTable <- ExampleAbeilleReconstructed
delta_count <- DeltaCount(SequencingTable, ReconstructedTable)
print(head(delta_count))
}
