% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleAbeilleDataSet.R
\docType{data}
\name{ExampleAbeilleDataSet}
\alias{ExampleAbeilleDataSet}
\title{Example of dataset usable in ABEILLE}
\format{
A data frame with transcripts in rows and samples in columns
}
\source{
Genotype-Tissue Expression, RNA gene read counts. \url{https://gtexportal.org/home/datasets}
}
\usage{
ExampleAbeilleDataSet
}
\description{
Random selected 14000 transcripts over 128 samples from the GTEx database
}
\examples{
\dontrun{
 ExampleAbeilleDataSet
}
}
\keyword{datasets}
