% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleAbeilleReconstructed.R
\docType{data}
\name{ExampleAbeilleReconstructed}
\alias{ExampleAbeilleReconstructed}
\title{Example of a reconstructed dataset usable in ABEILLE}
\format{
A data frame with transcripts in rows and samples in columns
}
\source{
Genotype-Tissue Expression, RNA gene read counts. \url{https://gtexportal.org/home/datasets}
}
\usage{
ExampleAbeilleReconstructed
}
\description{
Reconstruction of the randomly selected 14000 transcripts over 128 samples from the GTEx database
}
\examples{
\dontrun{
 ExampleAbeilleReconstructed
}
}
\keyword{datasets}
