% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneDisp.R
\name{GeneDisp}
\alias{GeneDisp}
\title{Estimate the gene dispersion}
\usage{
GeneDisp(SequencingTable)
}
\arguments{
\item{SequencingTable}{a read counts table with the transcripts in row and the samples in column.}
}
\value{
Vector of each transcripts gene dispersion.
}
\description{
Estimate the gene dispersion of each transcripts by using EdgeR estimateDisp function, then scale it to match OUTRIDER gene dispersion estimation.
}
\examples{
SequencingTable <- ExampleAbeilleDataSet
genedisp <- GeneDisp(SequencingTable)
print(head(genedisp))
}
