% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyAGE.R
\name{IdentifyAGE}
\alias{IdentifyAGE}
\title{Function to identify AGE}
\usage{
IdentifyAGE(
  divergence_score,
  delta_count,
  SequencingTable,
  ReconstructedTable,
  decision_tree = my_tree
)
}
\arguments{
\item{divergence_score}{data frame of the same size as the initial dataset. The divergence score can be computed with the function DivergenceScore.}

\item{delta_count}{data frame of the same size as the initial dataset. The delta count can be computed with the function DeltaCount.}

\item{SequencingTable}{a read counts table with the transcripts in row and the samples in column.}

\item{ReconstructedTable}{the reconstructed table generated by an autoencoder.}

\item{decision_tree}{function containing a custom decision tree. Check the help of the function my_tree for more informations.}
}
\value{
Data frame containing only rows predicted as AGE.
}
\description{
Using successively linear regression and decision tree, this function is able to pick AGE from two metric dataset. Originally built to work with the divergence score and the delta count, it is possible to feed the function with other metrics. The decision tree can be also customized.
}
\examples{
SequencingTable <- ExampleAbeilleDataSet
ReconstructedTable <- ExampleAbeilleReconstructed
divergence_score <- DivergenceScore(SequencingTable,ReconstructedTable)
delta_count <- DeltaCount(SequencingTable,ReconstructedTable)
ages <- IdentifyAGE(divergence_score,delta_count,SequencingTable, ReconstructedTable,
decision_tree = my_tree)
print(head(ages))
}
