% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2FC.R
\name{L2FC}
\alias{L2FC}
\title{Compute the Log2 fold-change}
\usage{
L2FC(SequencingTable, ReconstructedTable)
}
\arguments{
\item{SequencingTable}{a read counts table with the transcripts in row and the samples in column.}

\item{ReconstructedTable}{the reconstructed table generate by an autoencoder.}
}
\value{
All log2 fold-change in a matrix. The computation is the following one:
\ifelse{html}{
    \out{Δ<sub>ij</sub> = log2(k<sub>ij</sub> / μ<sub>i</sub><sup>k&#770)}}{
    \deqn{Δ_{ij} = log2(k_{ij} / μ&hat_{ij}^{k&hat})}}.
More detail in the ABEILLE paper.
}
\description{
Metric used to find outliers, this definition of the log2 fold-change is not the classic one.
}
\examples{
SequencingTable <- ExampleAbeilleDataSet
ReconstructedTable <- ExampleAbeilleReconstructed
l2fc <- L2FC(SequencingTable, ReconstructedTable)
print(head(l2fc))
}
