% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PickOutliers.R
\name{PickOutliers}
\alias{PickOutliers}
\title{Function to pick outliers}
\usage{
PickOutliers(
  divergence_score,
  delta_count,
  SequencingTable,
  ReconstructedTable,
  decision_tree = my_tree
)
}
\arguments{
\item{divergence_score}{data frame of the same size as the initial dataset. The divergence score can be compute with the function divergence_score.}

\item{delta_count}{data frame of the same size as the initial dataset. The log2 fold-change can be compute with the function delta_count.}

\item{SequencingTable}{a read counts table with the transcripts in row and the samples in column.}

\item{ReconstructedTable}{the reconstructed table generate by an autoencoder.}

\item{decision_tree}{function containing a custom decision tree. Check the help of the function my_tree for more information.}
}
\value{
Data frame containing only rows predict as outliers.
}
\description{
Successively using linear regression and decision tree this function is able to pick outliers from two metric dataset. originally built to work a the z-score and the log2 fold-change it is possible to feed the function with others metrics. The decision tree can be also custom.
}
\examples{
SequencingTable <- ExampleAbeilleDataSet
ReconstructedTable <- ExampleAbeilleReconstructed
divergence_score <- divergence_score(SequencingTable,ReconstructedTable)
delta_count <- delta_count(SequencingTable,ReconstructedTable)
genedisp <- GeneDisp(SequencingTable)
outliers <- PickOutliers(divergence_score,delta_count,SequencingTable, ReconstructedTable,
decision_tree = my_tree)
print(head(outliers))
}
