% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveZeroCounts.R
\name{RemoveZeroCounts}
\alias{RemoveZeroCounts}
\title{Remove zero counts}
\usage{
RemoveZeroCounts(SequencingTable)
}
\arguments{
\item{SequencingTable}{a read counts table with the transcripts in row and the samples in column.}
}
\value{
The table with the 0 count row removed and a print to indicate the number of transcript deleted.
}
\description{
Function to remove all transcripts with 0 count.
}
\examples{
SequencingTable <- ExampleAbeilleDataSet
WithoutZeroCounts <- RemoveZeroCounts(SequencingTable)
print(head(WithoutZeroCounts))
}
