% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SizeFactors.R
\name{SizeFactors}
\alias{SizeFactors}
\title{Estimate the size factor}
\usage{
SizeFactors(SequencingTable, annotation_table)
}
\arguments{
\item{SequencingTable}{a read counts table with the transcripts in row and the samples in column.}

\item{annotation_table}{a table of containing characteristics on each samples.}
}
\value{
Vector where each elements is the size factor of a sample.
}
\description{
Estimate the size factor of each samples by using DESeq2 estimateSizeFactors function.
}
\examples{
anno_table <- data.frame(sampleID = colnames(ExampleAbeilleDataSet),
Sex = sample(c("Male", "Female"), 128, replace = TRUE),
Age = sample(1:89, 128, replace = TRUE), Cohort = rep("GTEx database",128))
SequencingTable <- ExampleAbeilleDataSet
sizefactor <- SizeFactors(ExampleAbeilleDataSet, anno_table)
print(head(sizefactor))
}
