% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsPred.R
\name{StatsPred}
\alias{StatsPred}
\title{Display statistics on the error between of the input and the output of an autoencoder}
\usage{
StatsPred(SequencingTable, ReconstructedTable)
}
\arguments{
\item{SequencingTable}{a read counts table with the transcripts in row and the samples in column.}

\item{ReconstructedTable}{the reconstructed table generate by an autoencoder.}
}
\value{
Print in the console these statistics.
}
\description{
Print in the console different statistics of the absolute error between the two datasets provided.
}
\examples{
SequencingTable <- ExampleAbeilleDataSet
ReconstructedTable <- ExampleAbeilleReconstructed
StatsPred(SequencingTable,ReconstructedTable)
}
