% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_tree.R
\name{my_tree}
\alias{my_tree}
\title{Decision tree to choose AGEs}
\usage{
my_tree(linear_reg_data)
}
\arguments{
\item{linear_reg_data}{data of the linear regression.}
}
\value{
Boolean vector. This function is build to feed IdentifyAGE function, but you can build your own.
}
\description{
The decision tree to pick AGEs is a function using filters on different variables of the linear regression. This one was created by using CART on OUTRIDER outliers on the GTEx database. It is possible to use a custom decision tree by chaining ifelse conditions. Here are all variables usable:
\itemize{
 \item{"value"}{ the original value.}
 \item{"reconstruction"}{ the value reconstructed.}
 \item{"divergence_score"}{ see DivergenceScore function.}
 \item{"delta_count"}{ see DeltaCount function.}
 \item{"typeerror"}{ standardized residual. For more check rstandard function.}
 \item{"cooksD"}{ cook's distance. For more check cooks.distance function.}
 \item{"hat"}{ projection matrix. For more check hatvalues function.}
 \item{"dfbetas_intercept and dfbetas_var"}{ influence measures on the linear regression. For more check dfbetas function.}
}
}
\examples{
\dontrun{
 #The decision tree used:
 bool_vector <- ifelse(linear_reg_data['dfbetas_var'] >= 1.1,
 ifelse(linear_reg_data['hat'] >= 0.2, TRUE,
 ifelse(linear_reg_data['typeerror'] < 2.4, TRUE,
 ifelse(linear_reg_data['delta_count'] >= 2.4,
 ifelse(linear_reg_data['divergence_score'] >= 6.6, TRUE, FALSE), FALSE))),
 ifelse(linear_reg_data['divergence_score'] >=11, TRUE, FALSE))

 #A simple tree possible:
 bool_vector <- ifelse(linear_reg_data['cooksD'] >= 5, TRUE, FALSE)
}
}
