% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PickOutliers.R
\name{PickOutliers}
\alias{PickOutliers}
\title{Function to pick outliers}
\usage{
PickOutliers(
  zscore,
  l2fc,
  SequencingTable,
  ReconstructedTable,
  decision_tree = my_tree,
  nb_pval = FALSE
)
}
\arguments{
\item{zscore}{data frame of the same size as the initial dataset. The z-score can be compute with the function Zscore.}

\item{l2fc}{data frame of the same size as the initial dataset. The log2 fold-change can be compute with the function L2FC.}

\item{SequencingTable}{a read counts table with the transcripts in row and the samples in column.}

\item{ReconstructedTable}{the reconstructed table generate by an autoencoder.}

\item{decision_tree}{function containing a custom decision tree. Check the help of the function my_tree for more information.}

\item{nb_pval}{argument allowing to have an additional column in the result. The column is an estimation of the P-value using negative binomial distribution.}
}
\value{
Data frame containing only rows predict as outliers.
}
\description{
Successively using linear regression and decision tree this function is able to pick outliers from two metric dataset. originally built to work a the z-score and the log2 fold-change it is possible to feed the function with others metrics. The decision tree can be also custom.
}
\examples{
SequencingTable <- ExampleAbeilleDataSet
ReconstructedTable <- ExampleAbeilleReconstructed
zscore <- Zscore(SequencingTable,ReconstructedTable)
l2fc <- L2FC(SequencingTable,ReconstructedTable)
genedisp <- GeneDisp(SequencingTable)
outliers <- PickOutliers(zscore,l2fc,SequencingTable, ReconstructedTable,
decision_tree = my_tree, nb_pval=genedisp)
print(head(outliers))
}
